Added Game Over function and no monsters left check for game over.
Added a check everytime the user goes to the main menu. If they have no
monsters it check if the users gold is more than or equal to the
cheapest monster in the shop that day. Added a Game over function which
shows a game over sequence based on how the game ends.