Added a score tracker into the game.
Whenever a monster is killed, a function in Game Environment takes a
constant amount of points multiplied by the monster difficulty amount.
When the game ends the final score is the tracked points plus the user
gold left. Also added a monster kill tracked shown at the end of the
game.