Made a Main class to access the Application and updated README
Made a Main class as an entry point into the application. This is done
to facilitate the building of the GUI part of the program and so both
the CLA and the GUI are accessable from the command line. I updated the
readme with basic instructions on how to run the program both ways and
how to compile the source code.